Pongus and Snake
Pongus.html is a pong clone with some custom play rules and primitive
parallel game logic and rendering.
Snake.html is a snake clone with some custom rendering on the snake.
Snake was completed in one day as a prototyping example. Pongus was
created across three or so days as a somewhat more indepth experiment.